# Add 10 homework - program optimization